added people list container to a separate file with equatable
